refine

+++ oneflow/core/functional/impl/array_functor.cpp
@@ -2130,13 +2130,13 @@ class UnbindFunctor {
  public:
   UnbindFunctor() {}
   Maybe<TensorTuple> operator()(const std::shared_ptr<one::Tensor>& x, const int32_t& dim) const {
-    int64_t axis = dim;
+    int32_t axis = dim;
     const int32_t ndim = x->ndim();
     if (axis < 0) { axis += ndim; }
     CHECK_OR_RETURN((dim >= -ndim) && (dim < ndim))
         << "Dimension out of range (expected to be in range of [" << -ndim << "," << ndim - 1
         << "], but got " << dim << ")";
-    int64_t dim_size = x->shape()->At(axis);
+    int32_t dim_size = x->shape()->At(axis);
     TensorTuple unbinds(dim_size);
     for (int i = 0; i < dim_size; ++i) { unbinds[i] = JUST(Select(x, axis, i)); }
     return unbinds;
